abstract fade in text component
